test(client): fix unit tests
